NWBot/pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "discord-py>=2.6.4",
    "python-dotenv>=1.1.1",
]
